# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grocery Delivery API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface UploadResponse {
  url: string;
}

export interface SignupInput {
  name: string;
  phone: string;
  password: string;
  address?: string;
  profileImage?: string;
  latitude?: number;
  longitude?: number;
}

export interface LoginInput {
  phone: string;
  password: string;
}

export interface LocationInput {
  address?: string;
  latitude: number;
  longitude: number;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  phone: string;
  address?: string | null;
  profileImage?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  role: UserRole;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  nameAr: string;
  icon: string;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  nameAr: string;
  icon: string;
}

export type ProductUnit = (typeof ProductUnit)[keyof typeof ProductUnit];

export const ProductUnit = {
  kg: "kg",
  piece: "piece",
  bundle: "bundle",
} as const;

export interface Product {
  id: number;
  name: string;
  nameAr: string;
  description?: string | null;
  descriptionAr?: string | null;
  price: number;
  unit: ProductUnit;
  image?: string | null;
  categoryId?: number | null;
  category?: Category | null;
  featured: boolean;
  inStock: boolean;
  createdAt: string;
}

export type ProductInputUnit =
  (typeof ProductInputUnit)[keyof typeof ProductInputUnit];

export const ProductInputUnit = {
  kg: "kg",
  piece: "piece",
  bundle: "bundle",
} as const;

export interface ProductInput {
  name: string;
  nameAr: string;
  description?: string;
  descriptionAr?: string;
  price: number;
  unit: ProductInputUnit;
  image?: string;
  categoryId?: number;
  featured?: boolean;
  inStock?: boolean;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productNameAr: string;
  quantity: number;
  unit: string;
  price: number;
  subtotal: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  waiting: "waiting",
  accepted: "accepted",
  rejected: "rejected",
  preparing: "preparing",
  with_delivery: "with_delivery",
  completed: "completed",
} as const;

export interface DeliveryPerson {
  id: number;
  name: string;
  phone: string;
  active: boolean;
  createdAt: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  deliveryAddress?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  notes?: string | null;
  status: OrderStatus;
  totalPrice: number;
  items: OrderItem[];
  deliveryPersonId?: number | null;
  deliveryPerson?: DeliveryPerson | null;
  userId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface CreateOrderInput {
  customerName: string;
  customerPhone: string;
  deliveryAddress?: string;
  latitude?: number;
  longitude?: number;
  notes?: string;
  items: OrderItemInput[];
}

export type UpdateOrderStatusInputStatus =
  (typeof UpdateOrderStatusInputStatus)[keyof typeof UpdateOrderStatusInputStatus];

export const UpdateOrderStatusInputStatus = {
  accepted: "accepted",
  rejected: "rejected",
  preparing: "preparing",
  with_delivery: "with_delivery",
  completed: "completed",
} as const;

export interface UpdateOrderStatusInput {
  status: UpdateOrderStatusInputStatus;
}

export interface AssignDeliveryInput {
  deliveryPersonId: number;
}

export interface AssignDeliveryResponse {
  order: Order;
  whatsappSent: boolean;
  whatsappMessage?: string;
}

export interface DeliveryPersonInput {
  name: string;
  phone: string;
  active?: boolean;
}

export interface AdminStats {
  totalOrders: number;
  pendingOrders: number;
  completedOrders: number;
  totalProducts: number;
  totalCustomers: number;
  revenue: number;
}

export type GetProductsParams = {
  search?: string;
  categoryId?: number;
  featured?: boolean;
};

export type AdminGetOrdersParams = {
  status?: string;
};

export type UploadImageBody = {
  file?: Blob;
};