# lib/api-zod/src/generated/types/assignDeliveryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grocery Delivery API
 * OpenAPI spec version: 1.0.0
 */
import type { Order } from "./order";

export interface AssignDeliveryResponse {
  order: Order;
  whatsappSent: boolean;
  whatsappMessage?: string;
}