# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Grocery Delivery API
 * OpenAPI spec version: 1.0.0
 */
import type { DeliveryPerson } from "./deliveryPerson";
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  deliveryAddress?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  notes?: string | null;
  status: OrderStatus;
  totalPrice: number;
  items: OrderItem[];
  deliveryPersonId?: number | null;
  deliveryPerson?: DeliveryPerson | null;
  userId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}